create dictionary to store input arguments
This change takes the input argument from the user and passes it to
subfunction that creates a dictionary var called "inputs". This is
then read by subsequent functions to minimize # of separate variables.
To make this work, the conditional statement in the main script is
altered and a new subfunction is introduced. Other variable names are
changed to remove "mean" and use "statistic" or "output" as needed.

Signed-off-by: CMcC <carolyn.mcclaskey@gmail.com>