name of output CSV changed to include statistic
These changes modify how the output .csv file is constructed and
named, mostly by adding the exact input argument that the user
provided, and by using the same timestamp as was reported in the
terminal at the start of the program. These changes thus require the
addition of a subfunction that writes/formats the timestamp.

Signed-off-by: CMcC <carolyn.mcclaskey@gmail.com>