retain exact contents of input_file in output df
These changes create a new column "file" in the main datafram which
is separate from "input_file", for the purpose of keeping "input_file"
unchanged from what it was in the original input csv file. This is to
allow the user to check that the input_file column was correctly parsed
in the case of SPM syntax, and to verify that the file that was loaded
was the correct one. To do this, an extra input is used for the .nii
functions.

Signed-off-by: CMcC <carolyn.mcclaskey@gmail.com>