preserve input df in output df
This change modifies the final dataframe so that it's order is always
preserved and it contains all columns from the input dataframe. To do
this, merge is needed on 'index', and the indices need to be reset both
before and after merging. String in `input_file` are also stripped, just
in case.

Signed-off-by: CMcC <carolyn.mcclaskey@gmail.com>